Parse "String/Scientific" through newtype QuotedScientific
Many exchanges define amounts and prices as a quoted floating
point number (essentially a decimal), e.g. "1.98271398273"
which avoids the issue of being parsed as an imprecise Float/Double.

Instead of manually parsing this as a String, and then converting
to a Scientific afterwards, create a newtype "QuotedScientific" that
does this in one go.

Applies to Binance/Bitfinex/Bitstamp/Coinbase