Bitfinex: fix parsing of trading pair codes
Previously it was assumed that Bitfinex trading pair codes
were always six characters in length, and the first three
characters were the base currency and the following three
characters the quote currency. This was assumed to be the
case because all data fit this format (the trading pair
code string format is not documented).

Recent changes added trading pair codes whose base and
quote currency are separated by a colon. Add this option
to the parser.